+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Master (will become release 2.8)
 
+* Add an implementation of the Nédélec element of the first kind,
+  as introduced in "Nédélec, Mixed finite elements in R^3, 1980,
+  DOI: http://dx.doi.org/10.1007/BF01396415".
+  Only the first-order case for triangles is implemented.
+
 * Fix a bug in a shape function of the second-order Lagrange element
   on the three-dimensional pyramid.